refactor(admin/club): FSD 구조 변경 및 schemas/types 분리
