+++ apps/admin/src/features/club/hooks/index.ts
@@ -3,3 +3,4 @@ export * from './useCreateClub'
 export * from './useUpdateClub'
 export * from './useDeleteClub'
 export * from './useImageUpload'
+export * from './useEditTable'
