+++ apps/admin/src/features/club/schemas/index.ts
@@ -1,0 +1,4 @@
+export {
+  $ClubDetailResponse,
+  $ClubListResponse,
+} from '~/apis/community/requests/schemas.gen'
