+++ apps/admin/src/features/club/types/index.ts
@@ -1,0 +1,5 @@
+
+export type {
+  ClubDetailResponse,
+  ClubListResponse
+} from '~/apis/community/requests/types.gen'
